// src/lib/utils/supabase/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
      dream_analyses: {
        Row: {
          color_symbolism: string | null
          created_at: string | null
          gematria_interpretation: string | null
          id: string
          original_dream: string
          title: string
          updated_at: string | null
          user_id: string | null
        }
        Insert: {
          color_symbolism?: string | null
          created_at?: string | null
          gematria_interpretation?: string | null
          id?: string
          original_dream: string
          title: string
          updated_at?: string | null
          user_id?: string | null
        }
        Update: {
          color_symbolism?: string | null
          created_at?: string | null
          gematria_interpretation?: string | null
          id?: string
          original_dream?: string
          title?: string
          updated_at?: string | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "dream_analyses_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
      dream_entries: {
        Row: {
          analysis: Json
          created_at: string
          dream_analysis_id: string | null
          id: number
          user_id: string
        }
        Insert: {
          analysis: Json
          created_at?: string
          dream_analysis_id?: string | null
          id?: number
          user_id: string
        }
        Update: {
          analysis?: Json
          created_at?: string
          dream_analysis_id?: string | null
          id?: number
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "dream_entries_dream_analysis_id_fkey"
            columns: ["dream_analysis_id"]
            isOneToOne: false
            referencedRelation: "dream_analyses"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "dream_entries_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      interpretation_elements: {
        Row: {
          content: string
          dream_analysis_id: string | null
          id: string
          is_popover: boolean
          order_index: number
          popover_content: string | null
        }
        Insert: {
          content: string
          dream_analysis_id?: string | null
          id?: string
          is_popover: boolean
          order_index: number
          popover_content?: string | null
        }
        Update: {
          content?: string
          dream_analysis_id?: string | null
          id?: string
          is_popover?: boolean
          order_index?: number
          popover_content?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "interpretation_elements_dream_analysis_id_fkey"
            columns: ["dream_analysis_id"]
            isOneToOne: false
            referencedRelation: "dream_analyses"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      tags: {
        Row: {
          id: string
          name: string
        }
        Insert: {
          id?: string
          name: string
        }
        Update: {
          id?: string
          name?: string
        }
        Relationships: []
      }
// ...
      verses: {
        Row: {
          dream_analysis_id: string | null
          id: string
          reference: string
          text: string
        }
        Insert: {
          dream_analysis_id?: string | null
          id?: string
          reference: string
          text: string
        }
        Update: {
          dream_analysis_id?: string | null
          id?: string
          reference?: string
          text?: string
        }
        Relationships: [
          {
            foreignKeyName: "verses_dream_analysis_id_fkey"
            columns: ["dream_analysis_id"]
            isOneToOne: false
            referencedRelation: "dream_analyses"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}